Work in progress. Refactoring the event progress so one library can use the functionality of all three of these programs
