+++ CMakeLists.txt
@@ -1,7 +1,28 @@
+# dump1090 - ADS-B software radio decoder
+#
+# Copyright (C) 2018 Paul Ciarlo <paul.ciarlo@gmail.com>
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 cmake_minimum_required(VERSION 3.11)
 
 project(lib1090)
 
+#for android
+#add_subdirectory(deps)
+
+#for desktop linux with libusb-1.0 installed
 add_library(usb-1.0 SHARED IMPORTED)
 set_target_properties(
 	usb-1.0
@@ -16,18 +37,13 @@ set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -O3 -g -DENABLE_RTLSDR -std=c11 -D_DEFAULT_S
 set(SDR_SOURCES sdr.c sdr_ifile.c sdr_rtlsdr.c)
 
 #for linux build
-
-#set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -D_DEFAULT_SOURCE -Wall -Werror -Wno-unknown-warning-option -Wno-format-truncation -fPIC -Wl,visibility=default -Wl,export-dynamic")
-
 set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -D_DEFAULT_SOURCE -Wall -Werror -Wno-unknown-warning-option -Wno-format-truncation")
-#set(CMAKE_LD_FLAGS "${CMAKE_LD_FLAGS} -lrt -Wl,-soname,lib1090.so.0")
 set(COMPAT_SOURCES)
 set(EXTRA_LIBS "-lm -lncurses -lpthread")
 
 #for darwin build
 #set(COMPAT_SOURCES compat/clock_nanosleep/clock_nanosleep.c)
 #set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -DMISSING_NANOSLEEP -DCLOCKID_T)
-#set(CMAKE_LD_FLAGS "${CMAKE_LD_FLAGS}")
 #set(EXTRA_LIBS "-lm -lncurses -lpthread")
 
 add_library(
@@ -51,6 +67,10 @@ add_library(
 	stats.c
 	track.c
 	util.c
+	
+	dump1090.c
+	faup1090.c
+	view1090.c
 )
 
 target_link_libraries(
@@ -61,10 +81,10 @@ target_link_libraries(
 	"-Wl,-soname,lib1090.so.0"
 )
 
-add_executable(dump1090 dump1090.c)
+add_executable(dump1090 dump1090-main.c)
 target_link_libraries(dump1090 1090 ${EXTRA_LIBS})
-add_executable(faup1090 faup1090.c)
+add_executable(faup1090 faup1090-main.c)
 target_link_libraries(faup1090 1090 ${EXTRA_LIBS})
-add_executable(view1090 view1090.c)
+add_executable(view1090 view1090-main.c)
 target_link_libraries(view1090 1090 ${EXTRA_LIBS})
